requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
strands-agents[litellm]
bedrock-agentcore-starter-toolkit
streamlit>=1.49.1